Enable USB gadget support
The Duckbill now appears on a USB host system as network interface
which is bridged to the ethernet port (and on Duckbill SPI variants
to the powerline interface, too).

Additionally, a console is provided via USB. This way it's no
longer necessary to use a USB-to-serial adapter to connect to the
debug uart or fiddling with networking stuff like SSH/telnet, but
having access to a Linux console directly.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>